+++ src/modules/skyline/skyline.module.ts
@@ -1,5 +1,0 @@
-import { Module } from '@nestjs/common'
-import { SkylineController } from './skyline.controller'
-
-@Module({ controllers: [SkylineController] })
-export class SkylineModule {}